Move React Static Components into their old Folder
