+++ frontend/src/components/static/About.jsx
