+++ frontend/src/components/static/BooleanSearch.jsx
