+++ frontend/src/components/static/FinancialAssistance.jsx
