+++ frontend/src/components/static/JobHuntCompanies.jsx
