+++ frontend/src/components/static/JobHuntTips.jsx
