+++ frontend/src/components/static/ReleaseHistory.jsx
