+++ frontend/src/components/static/Secret.jsx
